# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HackR API",
        "description": "Here is my Swagger documentation for my HackR API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/images/random": {
            "get": {
                "tags": [
                    "Fonctionnalités"
                ],
                "summary": "Fetch a random image",
                "description": "Fetches a random image from an external service and returns it.",
                "operationId": "fetchRandomImage",
                "responses": {
                    "200": {
                        "description": "Random image retrieved successfully",
                        "content": {
                            "image/jpeg": {
                                "schema": [
                                    {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "error": "You are not authenticated."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - User does not have the necessary permissions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "error": "Vous n'avez pas le droit pour faire cela."
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to fetch the image",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "error": "Failed to fetch the image."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}